requirements.txt
fastapi==0.124.4
uvicorn[standard]==0.38.0
httpx==0.28.1
authlib==1.6.6
python-jose==3.3.0
pydantic==2.12.5
pydantic-settings==2.12.0
redis==6.4.0
starsessions==2.2.1
python-multipart==0.0.20
jinja2==3.1.3
pyyaml==6.0.1
loguru==0.7.2
python-jose[cryptography]==3.3.0
